fix(test): fix validator_switch_key_quick flakiness (#14741)
In the test, each shard has only one chunk producer. At some point we
stop the original chunk producer and start it on a failover node
instead. During the failover procedure, it may happen that the original
machine produces its final chunk, but the new machine is expected to
have the necessary transition data to produce the state witness for that
chunk.

There is a debug assertion that causes the new machine to fail in this
scenario:

https://github.com/near/nearcore/blob/d178e1830b062b407c270e8f8045753fd41cd081/chain/chain/src/stateless_validation/state_witness.rs#L219

Disabling this assertion makes the test pass, but it seemed more
appropriate to ensure that transition data is saved when the node is a
shadow validator, since a shadow validator is intended to serve as a
failover node.