feat(mirror): make MOCKNET_PROJECT and MOCKNET_STORE_PATH configurable (#14752)
The benchmarking scripts currently use hardcoded GCP project and mocknet
storage bucket.
Let's make them configurable using environment variables to make it
possible to run the benchmarks on other GCP accounts.

For neard-runner the `mocknet_storage_path` had to be added to the
config in order to upload it to the node together with neard-runner.