refactor(spice): separate constructor for spice chunk headers (#14740)
Using separate constructors allows to get rid of
`into_spice_chunk_execution_header` conversion helpers. It still leaves
genesis chunk to be a non-spice chunk for now.

This also allows to avoid passing defaults to the constructor in case of
spice which would not be used and pushes the check of whether spice is
on up the stack.

Refs https://github.com/near/nearcore/issues/14489