fix: fail block header validation early if epoch is unknown (#14757)
Right now we validate latest protocol version of a block header only if
we can retrieve epoch protocol version for block's epoch. IIUC this may
fail legitimately (not leading to conclusion that header is invalid)
only when we don't know the epoch. If that is the case however
validation would fail later on anyway when provenance isn't produced
(for example when calling in get_epoch_block_approvers_ordered) and when
it is produced it seems a reasonable assumption that we should have
information about the epoch).

Making the check fail early makes it easier to see that
epoch_protocol_version can be used for feature checks. Otherwise it's
very confusing and hard to tell in which cases it may not be available
for a valid header.

I also moved the check below the validations of the epoch id itself so
that in case of invalid epoch we would get a better error.

If you know where we have related unit-tests - let me know and I will
add unit test for this behaviour.