feat(spice): move final places to using spice protocol version check (#14786)
This migrates the last places in which at this time it's feasible to use
protocol version instead of compilation flag.

Here is the summary of where we still use cargo feature:
- checks during gc: spice-only db columns are behind cargo feature;
- test-loop setup;
- saving all next block hashes: this is spice-only db columns which is
behind cargo feature;
- creating spice actors for node and starting of spice actors: we want
to enable this only in the release that would migrate to spice;
- state sync: we should disable this once we implement stat sync support
for spice;
- view client actor querying: we would need to implement support for
getting blocks by new references and decide what to do during
transitional time.

Let me know if you think some of them can be migrated to depend on
protocol version check right now as well.

Refs https://github.com/near/nearcore/issues/14489