feat(epoch-sync): enable GC and cold store loop copy for DBCol::BlockHeader (#14795)
This PR enables garbage collection & cold store loop copy for the block
header column.

This is a follow on PR https://github.com/near/nearcore/pull/14738
Note that in parallel we have PRs open for doing the DB migration to
1. Move all historic block headers from hot store to cold store
2. Delete all old block headers in hot store before the gc period.

Note that the GC and cold store loop copy is gated on the protocol
feature `ContinuousEpochSync` being enabled at the binary level. This is
because we first need to have the migration happened before enabling
this feature.