feat(epoch-sync): implement db migration for continuous epoch sync (#14749)
This is the implementation of the DB migration for **hot store only**

This is a follow up on PR https://github.com/near/nearcore/pull/14738

We do the following in the db migration steps
1. As per the continuous epoch sync design, we now store the compressed
epoch sync proof. The migration generates the latest epoch sync proof
and stores the compressed version of the proof
2. We delete all the old block headers behind `tail`.

The implementation of deleting old block headers is in two steps
- We first delete the whole `DBCol::BlockHeader`. This is relatively a
quick operation.
- Then we iterate through all blocks from `tail` to
`latest_known_height` and generate the block header from the stored
block.

Note: I had originally tried implementing iterating from
`genesis_height` to `tail` and deleting all block headers but this was
prohibitively slow, requiring ~50 hrs for the migration to complete.

The current approach completes the migration in a bit over 2.5 minutes
on mainnet.

Tested the changes by deploying a mainnet node and running the
migration.