feat(network): enable sending `PartialEncodedChunkForward` over T1 (#14805)
We allowed receiving `PartialEncodedChunkForward` over T1 in 2.10. Since
protocol upgrade has already happened we expect the whole network to
adopt that. It means that it is now safe to enable sending over T1.

Notes:
- It is safe to completely remove the variant from `T2MessageBody` since
we don't send `RoutedMessage::V3` over the wire yet (we always receive
V1 and then convert it using `upgrade_to_v3`).
- With this change the node can still communicate with a node that sends
`PartialEncodedChunkForward` over T2 because we send `RoutedMessageBody`
over the wire. So on the receiving side it will be converted to
`T2MessageBody::PartialEncodedChunkForward` which is explicitly allowed
to be received over T1.

This is a followup to #14743.
Closes #14750.