feat(epoch-sync): serve epoch sync proof directly from store when continuous-epoch-sync is enabled (#14810)
When `ProtocolFeature::ContinuousEpochSync` is enabled, we automatically
update the epoch sync at the beginning of every epoch. This is handled
as part of PR https://github.com/near/nearcore/pull/14806

This means we for any incoming `EpochSyncRequestMessage` request, we can
simply serve it from store.

This PR additionally caches the compressed epoch sync proof at the store
layer. Note that we still need to clone to value for each request (which
was happening with the current in-memory cache anyway) but at least we
don't have to hit RocksDB store every time.